extract code section
