+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"H:\\programowanie\\102_JS_editor\\js-editor\\src\\index.tsx":"1","H:\\programowanie\\102_JS_editor\\js-editor\\src\\plugins\\unpkg-path-plugin.ts":"2","H:\\programowanie\\102_JS_editor\\js-editor\\src\\plugins\\fetch-plugin.ts":"3","H:\\programowanie\\102_JS_editor\\js-editor\\src\\components\\CodeEditor.tsx":"4","H:\\programowanie\\102_JS_editor\\js-editor\\src\\components\\Preview.tsx":"5"},{"size":1455,"mtime":1687891244655,"results":"6","hashOfConfig":"7"},{"size":838,"mtime":1687796140590,"results":"8","hashOfConfig":"7"},{"size":1863,"mtime":1687796140588,"results":"9","hashOfConfig":"7"},{"size":2467,"mtime":1687890625434,"results":"10","hashOfConfig":"7"},{"size":1034,"mtime":1687891579452,"results":"11","hashOfConfig":"7"},{"filePath":"12","messages":"13","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"19bcewr",{"filePath":"14","messages":"15","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"17","messages":"18","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"19","messages":"20","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"21","messages":"22","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"H:\\programowanie\\102_JS_editor\\js-editor\\src\\index.tsx",[],"H:\\programowanie\\102_JS_editor\\js-editor\\src\\plugins\\unpkg-path-plugin.ts",[],["23","24"],"H:\\programowanie\\102_JS_editor\\js-editor\\src\\plugins\\fetch-plugin.ts",[],"H:\\programowanie\\102_JS_editor\\js-editor\\src\\components\\CodeEditor.tsx",[],"H:\\programowanie\\102_JS_editor\\js-editor\\src\\components\\Preview.tsx",[],{"ruleId":"25","replacedBy":"26"},{"ruleId":"27","replacedBy":"28"},"no-native-reassign",["29"],"no-negated-in-lhs",["30"],"no-global-assign","no-unsafe-negation"]
+[{"H:\\programowanie\\102_JS_editor\\js-editor\\src\\index.tsx":"1","H:\\programowanie\\102_JS_editor\\js-editor\\src\\components\\CodeEditor.tsx":"2","H:\\programowanie\\102_JS_editor\\js-editor\\src\\components\\Preview.tsx":"3","H:\\programowanie\\102_JS_editor\\js-editor\\src\\bundler\\index.ts":"4","H:\\programowanie\\102_JS_editor\\js-editor\\src\\bundler\\plugins\\unpkg-path-plugin.ts":"5","H:\\programowanie\\102_JS_editor\\js-editor\\src\\bundler\\plugins\\fetch-plugin.ts":"6","H:\\programowanie\\102_JS_editor\\js-editor\\src\\components\\CodeSection.tsx":"7"},{"size":313,"mtime":1687903971030,"results":"8","hashOfConfig":"9"},{"size":2168,"mtime":1687904613724,"results":"10","hashOfConfig":"9"},{"size":1032,"mtime":1687906183574,"results":"11","hashOfConfig":"9"},{"size":760,"mtime":1687903741324,"results":"12","hashOfConfig":"9"},{"size":838,"mtime":1687901451159,"results":"13","hashOfConfig":"9"},{"size":1863,"mtime":1687796140588,"results":"14","hashOfConfig":"9"},{"size":999,"mtime":1687906409221,"results":"15","hashOfConfig":"9"},{"filePath":"16","messages":"17","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"19bcewr",{"filePath":"18","messages":"19","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"24"},{"filePath":"25","messages":"26","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"24"},{"filePath":"27","messages":"28","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"24"},{"filePath":"29","messages":"30","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"H:\\programowanie\\102_JS_editor\\js-editor\\src\\index.tsx",[],"H:\\programowanie\\102_JS_editor\\js-editor\\src\\components\\CodeEditor.tsx",[],"H:\\programowanie\\102_JS_editor\\js-editor\\src\\components\\Preview.tsx",[],"H:\\programowanie\\102_JS_editor\\js-editor\\src\\bundler\\index.ts",[],["31","32"],"H:\\programowanie\\102_JS_editor\\js-editor\\src\\bundler\\plugins\\unpkg-path-plugin.ts",[],"H:\\programowanie\\102_JS_editor\\js-editor\\src\\bundler\\plugins\\fetch-plugin.ts",[],"H:\\programowanie\\102_JS_editor\\js-editor\\src\\components\\CodeSection.tsx",[],{"ruleId":"33","replacedBy":"34"},{"ruleId":"35","replacedBy":"36"},"no-native-reassign",["37"],"no-negated-in-lhs",["38"],"no-global-assign","no-unsafe-negation"]
