+++ src/components/preview.css
@@ -1,0 +1,3 @@
+iframe {
+  background-color: white;
+}
