+++ src/index.tsx
@@ -1,51 +1,12 @@
 import ReactDOM from "react-dom";
-import { useState, useEffect, useRef } from "react";
-import * as esbuild from "esbuild-wasm";
-import { unpkgPathPlugin } from "./plugins/unpkg-path-plugin";
-import { fetchPlugin } from "./plugins/fetch-plugin";
-import Preview from "./components/Preview";
-
+import CodeSection from "./components/CodeSection";
 import "bulmaswatch/slate/bulmaswatch.min.css";
 
-import CodeEditor from "./components/CodeEditor";
-
 const App = () => {
-  const ref = useRef<any>();
-  const [code, setCode] = useState("");
-  const [input, setInput] = useState("");
-
-  const startService = async () => {
-    ref.current = await esbuild.startService({
-      worker: true,
-      wasmURL: "https://unpkg.com/esbuild-wasm@0.8.27/esbuild.wasm",
-    });
-  };
-
-  useEffect(() => {
-    startService();
-  }, []);
-
-  const onClick = async () => {
-    const result = await ref.current.build({
-      entryPoints: ["index.js"],
-      bundle: true,
-      write: false,
-      plugins: [unpkgPathPlugin(), fetchPlugin(input)],
-      define: {
-        "process.env.NODE_ENV": '"production"',
-        global: '"window"',
-      },
-    });
-    setCode(result.outputFiles[0].text);
-  };
-
   return (
     <div>
-      <CodeEditor initialValue={null} onChange={(val) => setInput(val)} />
-      <div>
-        <button onClick={onClick}>Submit</button>
-      </div>
-      <Preview code={code} />
+      <CodeSection />
+      <CodeSection />
     </div>
   );
 };
